Reverted accidental deletion

+++ src/user/find.cpp
@@ -41,7 +41,7 @@ size_t __stdcall find(const kiv_hal::TRegisters& regs) {
 		print_count_lines = true;
 	}
 
-	kiv_os_rtl::Open_File(filename, kiv_os::NFile_Attributes::System_File, kiv_os::NOpen_File::fmOpen_Always, file_handle);
+	kiv_os_rtl::Open_File(filename, static_cast<uint8_t>(kiv_os::NFile_Attributes::System_File), kiv_os::NOpen_File::fmOpen_Always, file_handle);
 	
 	if (file_handle == invalid_file_handle) {
 		kiv_os_rtl::Write_File(stdout_handle, ERROR_MSG_CANT_OPEN_FILE.data(), ERROR_MSG_CANT_OPEN_FILE.size(), chars_written);
